Add JSON mixin, fix list bug
